+++ src/private-key.ts
@@ -1,17 +1,15 @@
+import { BigInteger, KEYUTIL, KJUR, RSAKey, stob64, X509 } from 'jsrsasign';
+import { Mixin } from 'ts-mixer';
+
 import { Key } from './internal/key';
-import { delegate } from 'typescript-mix';
 import { LocalFileOpenTrait } from './internal/local-file-open-trait';
 import { PublicKey } from './public-key';
-import { BigInteger, KEYUTIL, KJUR, RSAKey, stob64, X509 } from 'jsrsasign';
 import { SignatureAlgorithm } from './signature-algorithm';
 import { PemExtractor } from './pem-extractor';
 import { KeyType } from './internal/key-type-enum';
 import { Certificate } from './certificate';
 
-export class PrivateKey extends Key {
-    @delegate(LocalFileOpenTrait.localFileOpen)
-    private static localFileOpen: (filename: string) => string;
-
+export class PrivateKey extends Mixin(Key, LocalFileOpenTrait) {
     /** string PEM contents of private key  */
     private readonly _pem: string;
 
@@ -24,8 +22,8 @@ export class PrivateKey extends Key {
     /**
      * Private key constructor
      *
-     * @param source can be a PKCS#8 DER, PKCS#8 PEM or PKCS#5 PEM
-     * @param passPhrase if empty asume unencrypted/plain private key
+     * @param source - can be a PKCS#8 DER, PKCS#8 PEM or PKCS#5 PEM
+     * @param passPhrase - if empty asume unencrypted/plain private key
      */
     constructor(source: string, passPhrase: string) {
         super({});
@@ -41,46 +39,34 @@ export class PrivateKey extends Key {
         }
         this._pem = pem;
         this._passPhrase = passPhrase;
-        this.dataRecord = this.callOnPrivateKey(
-            (privateKey: RSAKey | KJUR.crypto.DSA | KJUR.crypto.ECDSA): Record<string, unknown> => {
-                try {
-                    const privateKeyJWK = KEYUTIL.getJWKFromKey(privateKey);
-                    const { e, kty, n } = privateKeyJWK;
-                    const pubKeyJWK = { e, kty, n };
-                    const pubKey = KEYUTIL.getKey(pubKeyJWK);
-                    const pubKeyPEMText = KEYUTIL.getPEM(pubKey);
-                    const data: Record<string, unknown> = {};
-                    if (privateKey instanceof RSAKey) {
-                        data['bits'] = (privateKey as unknown as { n: BigInteger }).n.bitLength();
-                        data['key'] = pubKeyPEMText;
-                        data[KeyType.RSA] = privateKey;
-                        data['type'] = KeyType.RSA;
-                    } else if (privateKey instanceof KJUR.crypto.DSA) {
-                        data['bits'] = (privateKey as unknown as { p: BigInteger }).p.bitLength();
-                        data['key'] = pem;
-                        data[KeyType.DSA] = pubKey;
-                        data['type'] = KeyType.DSA;
-                    }
-                    return data;
-                } catch (e) {
-                    throw new Error('Cannot open private key');
-                }
-            }
-        );
+        this.dataRecord = this.callOnPrivateKey((privateKey): Record<string, unknown> => {
+            const data: Record<string, unknown> = {};
+            const privKeyJWK = KEYUTIL.getJWKFromKey(privateKey);
+            const { e, kty, n } = privKeyJWK;
+            const pubKey = KEYUTIL.getKey({ e, kty, n });
+
+            data['bits'] = (privateKey as unknown as { n: BigInteger }).n.bitLength();
+            data['key'] = KEYUTIL.getPEM(pubKey);
+            data[KeyType.RSA] = privateKey;
+            data['type'] = KeyType.RSA;
+
+            return data;
+        });
     }
 
     /**
      * Convert PKCS#8 DER to PKCS#8 PEM
      *
-     * @param contents
-     * @param isEncrypted
+     * @param contents -
+     * @param isEncrypted -
      */
     public static convertDerToPem(contents: string, isEncrypted: boolean): string {
         const privateKeyName = isEncrypted ? 'ENCRYPTED PRIVATE KEY' : 'PRIVATE KEY';
+
         return [
             `-----BEGIN ${privateKeyName}-----\n`,
             `${(stob64(contents).match(/.{1,64}/g) || []).join('\n')}\n`,
-            `-----END ${privateKeyName}-----`,
+            `-----END ${privateKeyName}-----`
         ].join('');
     }
 
@@ -88,8 +74,10 @@ export class PrivateKey extends Key {
      * Create a PrivateKey object by opening a local file
      * The content file can be a PKCS#8 DER, PKCS#8 PEM OR PKCS#5 PEM
      *
-     * @param filename
-     * @param passPhrase
+     * @param filename - file name to be read
+     * @param passPhrase - if file is encrypted
+     *
+     * This function only works in Node.js.
      */
     public static openFile(filename: string, passPhrase: string): PrivateKey {
         return new PrivateKey(PrivateKey.localFileOpen(filename), passPhrase);
@@ -107,22 +95,28 @@ export class PrivateKey extends Key {
         if (!this._publicKey) {
             this._publicKey = new PublicKey(this.publicKeyContents());
         }
+
         return this._publicKey;
     }
 
+    /**
+     * Sign string data by provider algorithm
+     *
+     * @param data - input data
+     * @param algorithm - algorithm to be used
+     * @returns Hexadecimal string signature
+     */
     public sign(data: string, algorithm: SignatureAlgorithm = SignatureAlgorithm.SHA256): string {
-        return this.callOnPrivateKey((privateKey: unknown) => {
-            let signature: string | null = null;
+        return this.callOnPrivateKey((privateKey) => {
             try {
-                signature = (privateKey as { sign(data: string, alg: string): string }).sign(data, algorithm);
+                const sig = new KJUR.crypto.Signature({ alg: algorithm });
+                sig.init(privateKey);
+                sig.updateString(data);
+
+                return sig.sign();
             } catch (e) {
                 throw new Error('Cannot sign data: empty signature');
             }
-            signature = `${signature}`;
-            if ('' == signature) {
-                throw new Error('Cannot sign data: empty signature');
-            }
-            return signature;
         });
     }
 
@@ -135,16 +129,23 @@ export class PrivateKey extends Key {
         const x = new X509();
         x.readCertPEM(certificate);
         const certPubKey = x.getPublicKey();
+
         return JSON.stringify(certPubKey) === JSON.stringify(pubKey);
     }
 
-    public callOnPrivateKey<T>(callableFunction: CallableFunction): T {
-        let privateKey: RSAKey | KJUR.crypto.DSA | KJUR.crypto.ECDSA | null = null;
+    public callOnPrivateKey<T>(callableFunction: (prv: RSAKey) => T): T {
+        let privateKey: RSAKey | KJUR.crypto.DSA | KJUR.crypto.ECDSA | undefined = undefined;
         try {
             privateKey = KEYUTIL.getKey(this._pem, this._passPhrase);
         } catch (e) {
-            if (e instanceof Error) throw new Error(`Cannot open private key: ${e.message}`);
+            throw new Error(`Cannot open private key: ${(e as Error).message}`);
         }
+
+        /* istanbul ignore next */
+        if (!privateKey || privateKey instanceof KJUR.crypto.ECDSA || privateKey instanceof KJUR.crypto.DSA) {
+            throw new Error(`Cannot open private key: type not supported only RSAKey is accepted`);
+        }
+
         return callableFunction(privateKey);
     }
 }