+++ tests/unit/internals/rfc4514.test.ts
@@ -1,4 +1,4 @@
-import { Rfc4514 } from '../../../src/internal/rfc4514';
+import { Rfc4514 } from '~/internal/rfc4514';
 
 describe('Rfc4514', () => {
     let object: Rfc4514;
@@ -18,8 +18,8 @@ describe('Rfc4514', () => {
         ['with inner #', 'foo#bar', 'foo#bar'],
         ['with trail #', 'foo bar#', 'foo bar#'],
         ['with =', '=foo=bar=', '\\3dfoo\\3dbar\\3d'],
-        ['complex : [# a=1,b>2,c<3 ]', '# a=1,b>2,c<3 ', '\\22 a\\3d1\\2cb\\3e2\\2cc\\3c3\\20'],
-    ])('escape %s', (name: string, subject: string, expected: string) => {
+        ['complex : [# a=1,b>2,c<3 ]', '# a=1,b>2,c<3 ', '\\22 a\\3d1\\2cb\\3e2\\2cc\\3c3\\20']
+    ])('escape %s', (_name: string, subject: string, expected: string) => {
         expect(object.escape(subject)).toBe(expected);
     });
 
@@ -28,13 +28,13 @@ describe('Rfc4514', () => {
             'foo': 'foo bar',
             '#foo': '#foo bar',
             'foo ': 'foo bar ',
-            'address': 'Street #1, Somewhere',
+            'address': 'Street #1, Somewhere'
         };
         const expected = [
             'foo=foo bar',
             '\\22foo=\\22foo bar',
             'foo\\20=foo bar\\20',
-            'address=Street #1\\2c Somewhere',
+            'address=Street #1\\2c Somewhere'
         ].join(',');
         expect(object.escapeRecord(subject)).toBe(expected);
     });