+++ tests/unit/serial-number.test.ts
@@ -1,4 +1,4 @@
-import { SerialNumber } from '../../src';
+import { SerialNumber } from '~/serial-number';
 
 describe('SerialNumber', () => {
     const SERIAL_HEXADECIMAL = '3330303031303030303030333030303233373038';
@@ -14,18 +14,30 @@ describe('SerialNumber', () => {
     });
 
     test('create hexadecimal empty', () => {
-        expect(() => SerialNumber.createFromHexadecimal('')).toThrow(SyntaxError);
-        expect(() => SerialNumber.createFromHexadecimal('')).toThrow('is empty');
+        const t = (): SerialNumber => {
+            return SerialNumber.createFromHexadecimal('');
+        };
+
+        expect(t).toThrow(Error);
+        expect(t).toThrow('is empty');
     });
 
     test('create hexadecimal invalid chars', () => {
-        expect(() => SerialNumber.createFromHexadecimal('0x001122x3')).toThrow(SyntaxError);
-        expect(() => SerialNumber.createFromHexadecimal('0x001122x3')).toThrow('contains invalid characters');
+        const t = (): SerialNumber => {
+            return SerialNumber.createFromHexadecimal('0x001122x3');
+        };
+
+        expect(t).toThrow(Error);
+        expect(t).toThrow('contains invalid characters');
     });
 
     test('create hexadecimal double prefix', () => {
-        expect(() => SerialNumber.createFromHexadecimal('0x0xFF')).toThrow(SyntaxError);
-        expect(() => SerialNumber.createFromHexadecimal('0x0xFF')).toThrow('contains invalid characters');
+        const t = (): SerialNumber => {
+            return SerialNumber.createFromHexadecimal('0x0xFF');
+        };
+
+        expect(t).toThrow(Error);
+        expect(t).toThrow('contains invalid characters');
     });
 
     test('create from decimal', () => {
@@ -33,7 +45,7 @@ describe('SerialNumber', () => {
         expect(serial.bytes()).toBe(SERIAL_BYTES);
     });
 
-    test('create form bytes', () => {
+    test('create from bytes', () => {
         const serial = SerialNumber.createFromBytes(SERIAL_BYTES);
         expect(serial.hexadecimal()).toBe(SERIAL_HEXADECIMAL);
     });